Extract NormalizeController from extension
